include alternative keypoint detectors
Currently it wont build on my machine due to SIFT and the nonfree opencv modules.
I think this is just my local build environment causing problems.
About to fix it, but for now I know things work if SIFT is not used and
and the line referencing it in matching2D_Student.cpp is commented out